Move some parts to native in T_T Pandoras Box

+++ Api/Game/Offsets/OffsetData.cs
@@ -5,13 +5,13 @@ namespace Api.Game.Offsets;
 public struct OffsetData
 {
     public string Name { get; }
-    public int Offset { get; }
-    public int TargetSize { get; }
+    public uint Offset { get; }
+    public uint TargetSize { get; }
 
-    public OffsetData(string name, int offset, Type type)
+    public OffsetData(string name, uint offset, Type type)
     {
         Name = name;
         Offset = offset;
-        TargetSize = Marshal.SizeOf(type);
+        TargetSize = (uint)Marshal.SizeOf(type);
     }
 }
